+++ R/plot.POMDP.R
@@ -16,9 +16,12 @@ plot_policy_graph <- function(x, belief = TRUE, legend = TRUE,
 }  
 
 
-.plot.igraph <- function(x, belief, legend, col, ...) {
-  pg<- policy_graph(x, belief = belief, col = col)
-  plot.igraph(pg, ...)
+.plot.igraph <- function(x, belief, legend, col, edge.curved = NULL, ...) {
+  pg <- policy_graph(x, belief = belief, col = col)
+  
+  if(is.null(edge.curved)) edge.curved <- .curve_multiple_directed(pg)
+  
+  plot.igraph(pg, edge.curved = edge.curved, ...)
   
   if(legend && belief && !is.null(V(pg)$pie)) {
     legend("topright", legend = x$model$states, title = "Belief", 
@@ -31,22 +34,21 @@ plot_policy_graph <- function(x, belief = TRUE, legend = TRUE,
 }
 
 ### fix the broken curve_multiple for directed graphs (igraph_1.2.2)
-# curve_multiple_fixed <- function(graph, start = 0.5) 
-# {
-#   el <-  as_edgelist(graph, names = FALSE)
-#   o <- apply(el, 1, order)[1,]
-#   el <- apply(el, 1, FUN = function(x) paste(sort(x), collapse = ":"))
-#   cu <- ave(rep(NA, length(el)), el, FUN = function(x) {
-#     if (length(x) == 1) {
-#       return(0)
-#     }
-#     else {
-#       return(seq(-start, start, length = length(x)))
-#     }
-#   }
-#   )
-#   
-#   cu[o==2] <- cu[o==2] * -1
-#   cu
-# }
+.curve_multiple_directed <- function(graph, start = 0.3) {
+   el <-  as_edgelist(graph, names = FALSE)
+   o <- apply(el, 1, order)[1,]
+   el <- apply(el, 1, FUN = function(x) paste(sort(x), collapse = ":"))
+   cu <- ave(rep(NA, length(el)), el, FUN = function(x) {
+     if (length(x) == 1) {
+       return(0)
+     }
+     else {
+       return(seq(-start, start, length = length(x)))
+     }
+   }
+   )
+   
+   cu[o==2] <- cu[o==2] * -1
+   cu
+ }