+++ R/policy_graph.R
@@ -4,15 +4,15 @@ policy_graph <- function(x, belief = TRUE, col = NULL) {
  
   .solved_POMDP(x) 
   
-  if(!x$solution$converged) warning("POMDP has not converged. The last epoch in the policy tree may not form a graph! Use with caution!")
- 
-  ## try to make a graph from a not converged policy
-  ## FIXME: This needs to be improved!
-  if(!is.data.frame(x$solution$pg) && nrow(x$solution$pg[[1]]) != nrow(x$solution$pg[[2]]))
-    stop("Number of nodes the of last two epoch does not agree! Cannot create graph!")
+  pg <- x$solution$pg[[1]]
   
-  ## producing the optimal policy graph
-  pg <- if(is.data.frame(x$solution$pg)) x$solution$pg else x$solution$pg[[1]]
+  ## FIXME: try to make a graph from a not converged policy
+  if(!x$solution$converged){ 
+    if(nrow(x$solution$pg[[1]]) != nrow(x$solution$pg[[2]]))
+      stop("Number of nodes the of last two epochs does not agree! Cannot create graph!")
+    
+    warning("POMDP has not converged. The last epoch in the policy tree may not form a graph! Use with caution!")
+  }
   
   # producing a list containing arcs
   l <- list()